Fix build flag with new interface
