+++ pkg/util/kubernetes/kubelet/kubelet.go
@@ -45,24 +45,6 @@ var (
 	kubeletExpVar       = expvar.NewInt("kubeletQueries")
 )
 
-// KubeUtilInterface defines the interface for kubelet api
-type KubeUtilInterface interface {
-	GetNodeInfo() (string, string, error)
-	GetNodename() (string, error)
-	GetLocalPodList() ([]*Pod, error)
-	ForceGetLocalPodList() ([]*Pod, error)
-	GetPodForContainerID(containerID string) (*Pod, error)
-	GetStatusForContainerID(pod *Pod, containerID string) (ContainerStatus, error)
-	GetPodFromUID(podUID string) (*Pod, error)
-	GetPodForEntityID(entityID string) (*Pod, error)
-	QueryKubelet(path string) ([]byte, int, error)
-	GetKubeletApiEndpoint() string
-	GetRawConnectionInfo() map[string]string
-	GetRawMetrics() ([]byte, error)
-	ListContainers() ([]*containers.Container, error)
-	UpdateContainerMetrics(ctrList []*containers.Container) error
-}
-
 // KubeUtil is a struct to hold the kubelet api url
 // Instantiate with GetKubeUtil
 type KubeUtil struct {
